Move FlowDocuments class to flow resource module
