+++ airslate/facades.py
@@ -5,10 +5,9 @@
 # For the full copyright and license information, please view
 # the LICENSE file that was distributed with this source code.
 
-"""Facade module for API resources."""
+"""Facades module for API resources."""
 
-from .resources.addons import FlowDocuments
-from .resources.slate import Tags
+from .resources import flow, slate
 
 
 class Flow:  # pylint: disable=too-few-public-methods
@@ -21,9 +20,9 @@ class Flow:  # pylint: disable=too-few-public-methods
 
     @property
     def documents(self):
-        """Getter for :class:`FlowDocuments`."""
+        """Facade for :class:`flow.Documents`."""
         if self._documents is None:
-            self._documents = FlowDocuments(self._client)
+            self._documents = flow.Documents(self._client)
         return self._documents
 
 
@@ -37,7 +36,7 @@ class Slate:  # pylint: disable=too-few-public-methods
 
     @property
     def tags(self):
-        """Getter for :class:`Tags`."""
+        """Facade for :class:`slate.Tags`."""
         if self._tags is None:
-            self._tags = Tags(self._client)
+            self._tags = slate.Tags(self._client)
         return self._tags
