+++ airslate/resources/addons.py
@@ -5,9 +5,8 @@
 # For the full copyright and license information, please view
 # the LICENSE file that was distributed with this source code.
 
-"""Addons module for airslate package."""
+"""Addons API resource module."""
 
-from airslate.entities.documents import Document
 from . import BaseResource
 
 
@@ -30,16 +29,3 @@ class Addons(BaseResource):
         }
 
         return self.client.post(url, data, headers=headers, full_response=True)
-
-
-class FlowDocuments(BaseResource):
-    """Represent Flow Documents resource."""
-
-    def collection(self, flow_id, **options):
-        """Get supported Documents for a given Flow."""
-        url = self.resolve_endpoint(
-            f'addons/slates/{flow_id}/documents'
-        )
-
-        response = self.client.get(url, full_response=True, **options)
-        return Document.from_collection(response)
