+++ airslate/resources/slate_addon.py
@@ -5,7 +5,7 @@
 # For the full copyright and license information, please view
 # the LICENSE file that was distributed with this source code.
 
-"""Slate Addon module for airslate package."""
+"""Slate Addon API resource module."""
 
 from airslate.entities.addons import SlateAddonFile
 from . import BaseResource
